auctionService configurado para cliente de token de identity server

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -6,6 +6,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -65,12 +66,13 @@ public class AuctionsController: ControllerBase
         return Ok(_mapper.Map<AuctionDto>(auction));
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
-        // TODO: Add current user as seller
-        auction.Seller = "test";
+        
+        auction.Seller = User.Identity.Name;
         
         _context.Auctions.Add(auction);
 
@@ -91,6 +93,7 @@ public class AuctionsController: ControllerBase
                         newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -103,7 +106,11 @@ public class AuctionsController: ControllerBase
             return NotFound();
         }
 
-        // TODO: Check seller == username
+        // Check seller == username
+        if(auction.Seller != User.Identity.Name)
+        {
+            return Forbid();
+        }
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -124,6 +131,7 @@ public class AuctionsController: ControllerBase
         return Ok();
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -134,7 +142,11 @@ public class AuctionsController: ControllerBase
             return NotFound();
         }
 
-        // TODO: Check seller == username
+        // Check seller == username
+        if(auction.Seller != User.Identity.Name)
+        {
+            return Forbid();
+        }
 
         _context.Auctions.Remove(auction);
         await _publishEndpoint.Publish(new AuctionDeleted { Id = id.ToString() });
